Refactoring renvoi des fonctionnalités de la carte dans le composant correspondant
