+++ src/pages/home/home.html
@@ -6,8 +6,4 @@
 
 <ion-content>
   <div id="map" #map class="map"></div>
-  <!--<div id="popup">-->
-    <!--<a href="#" id="popup-closer" class="ol-popup-closer"><ion-icon name="close-circle"></ion-icon></a>-->
-    <!--<div id="popup-content"></div>-->
-  <!--</div>-->
 </ion-content>
